infra/pyproject.toml
[project]
name = "turboflow-infra"
version = "0.1.0"
description = "TurboFlow Cloud — Phase 3 serverless infrastructure (CDK)"
requires-python = ">=3.10"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "aws-cdk.aws-bedrock-agentcore-alpha>=2.170.0a0",
    "constructs>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["turboflow_infra"]

[tool.ruff]
target-version = "py310"
line-length = 100
